fixed cards/components and added itinerary preview button
