+++ scripts/main.js
@@ -4,14 +4,13 @@ import { getAttractions } from "./attractions/AttractionProvider.js"
 import AttractionSelect from "./attractions/AttractionSelect.js"
 import { getEateries } from "./eateries/EateryProvider.js"
 import { eaterySelect } from "./eateries/EaterySelect.js"
-import  "./eateries/EateryCard.js"
+import  "./ItineraryPreview/ItineraryPreview.js"
 import "./weather/WeatherList.js"
-import "./attractions/AttractionCard.js"
 import { eateryDialogButton } from "./eateries/EateryDetailDialog.js"
-import "./parks/ParkCard.js"
 import { attractionDialogButton } from "./attractions/AttractionDetailDialog.js"
 import { renderItinerary } from "./Itinerary/ItineraryList.js"
 import { getItineraries } from "./Itinerary/ItineraryProvider.js"
+import { displayItineraryButton } from "./ItineraryPreview/SavedItineraryButton.js"
 
 getAttractions()
     .then(AttractionSelect)
@@ -23,3 +22,5 @@ getParks()
 getEateries()
     .then(eaterySelect)
     .then(eateryDialogButton)
+
+displayItineraryButton()
